test(server): split api-hardening coverage by route
