+++ packages/server/tests/auth-bot-passwords.test.ts
@@ -37,6 +37,63 @@ describe("auth bot password restrictions", () => {
     expect(created).toBeUndefined();
   });
 
+  it("stores reset tokens hashed and resolves hashed lookup", async () => {
+    db.prepare("INSERT INTO accounts (id, username, email, email_verified, is_bot) VALUES (?, ?, ?, 1, 0)")
+      .run("u2", "bob", "bob@example.com");
+    const app = makeApp(db);
+
+    await app.request("http://localhost/api/v1/auth/forgot-password", {
+      method: "POST",
+      headers: { "content-type": "application/json" },
+      body: JSON.stringify({ email: "bob@example.com" }),
+    });
+
+    const stored = db.prepare("SELECT token FROM password_reset_tokens WHERE account_id = ?").get("u2") as { token: string };
+    expect(stored.token).toMatch(/^[a-f0-9]{64}$/);
+
+    db.prepare("UPDATE password_reset_tokens SET token = ? WHERE account_id = ?").run(hashTokenSecret("known-token"), "u2");
+    const resetRes = await app.request("http://localhost/api/v1/auth/reset-password", {
+      method: "POST",
+      headers: { "content-type": "application/json" },
+      body: JSON.stringify({ token: "known-token", newPassword: "new-password-123" }),
+    });
+    expect(resetRes.status).toBe(200);
+  });
+
+  it("does not allow toggling isBot via auth endpoints", async () => {
+    db.prepare("INSERT INTO accounts (id, username, password_hash, email_verified, is_bot) VALUES (?, ?, ?, 1, 0)")
+      .run("u3", "carol", hashPassword("pw"));
+    const app = makeApp(db, { id: "u3", username: "carol" });
+
+    const registerRes = await app.request("http://localhost/api/v1/auth/register", {
+      method: "POST",
+      headers: { "content-type": "application/json" },
+      body: JSON.stringify({ username: "x", isBot: true }),
+    });
+    expect(registerRes.status).toBe(400);
+
+    const patchRes = await app.request("http://localhost/api/v1/auth/me", {
+      method: "PATCH",
+      headers: { "content-type": "application/json" },
+      body: JSON.stringify({ isBot: true, displayName: "Carol" }),
+    });
+    expect(patchRes.status).toBe(200);
+    const row = db.prepare("SELECT is_bot FROM accounts WHERE id = ?").get("u3") as { is_bot: number };
+    expect(row.is_bot).toBe(0);
+  });
+
+  it("returns 400 for malformed JSON on auth login route", async () => {
+    const app = makeApp(db);
+    db.prepare("INSERT INTO accounts (id, username, email_verified) VALUES (?, ?, 1)").run("u1", "alice");
+
+    const res = await app.request("http://localhost/api/v1/auth/login", {
+      method: "POST",
+      headers: { "content-type": "application/json" },
+      body: "{",
+    });
+    expect(res.status).toBe(400);
+  });
+
   it("creates normal user account when isBot is omitted", async () => {
     const app = makeApp(db);
     const res = await app.request("http://localhost/api/v1/auth/register", {
